Working on delete user account API
